romcal export: map celebration IDs to our own
... if possible. Otherwise the romcal IDs are shortenned and converted to
snake_case.

Note: romcal seems not to have General Roman Calendar up to date;
apart of that, deals incorrectly with some optional memorials,
merging them in a single memorial of multiple saints.

Solemnities and some feasts are missing in general.js and probably are
handled by some other part of the library, which makes the resulting
data file of little use with calendarium-romanum, which expects all
sanctorale celebrations to be defined in a sanctorale data file.

resolves #2